move tc to its own package and use net.Interface
